+++ app/src/main/java/com/example/budgettracker/analytics/IncomePieFragment.kt
@@ -71,7 +71,7 @@ class IncomePieFragment : Fragment() {
             for (i in 0 until resultList.size) {
                 resultList[i].color = customColors[i % customColors.size]
             }
-            resultList.sortByDescending { it.amount.toInt() }
+            resultList.sortByDescending { it.amount.toDouble() }
             binding.operationsRV.adapter = PieAdapter(resultList, findNavController(), viewModel)
             viewModel.analyzedCategoriesListIncome = resultList
             viewModel.analyzedMonthIndexIncome = selectedMonthIndex
@@ -87,7 +87,7 @@ class IncomePieFragment : Fragment() {
                 for (i in 0 until resultList.size) {
                     resultList[i].color = customColors[i]
                 }
-                resultList.sortByDescending { it.amount.toInt() }
+                resultList.sortByDescending { it.amount.toDouble() }
                 binding.operationsRV.adapter = PieAdapter(resultList, findNavController(), viewModel)
                 viewModel.analyzedCategoriesListIncome = resultList
                 viewModel.analyzedMonthIndexIncome = selectedMonthIndex
@@ -103,7 +103,7 @@ class IncomePieFragment : Fragment() {
                 for (i in 0 until resultList.size) {
                     resultList[i].color = customColors[i]
                 }
-                resultList.sortByDescending { it.amount.toInt() }
+                resultList.sortByDescending { it.amount.toDouble() }
                 binding.operationsRV.adapter = PieAdapter(resultList, findNavController(), viewModel)
                 viewModel.analyzedCategoriesListIncome = resultList
                 viewModel.analyzedMonthIndexIncome = selectedMonthIndex
@@ -117,7 +117,7 @@ class IncomePieFragment : Fragment() {
     private fun setupPie(selectedMonthIndex : Int, incomeDataList : ArrayList<ArrayList<OperationsData>>, customColors : IntArray) {
         val entries = ArrayList<PieEntry>()
         categoriesMap = mutableMapOf()
-        var totalForMonth = 0
+        var totalForMonth = 0.0
 
         val calendar = Calendar.getInstance()
         if (incomeDataList.isNotEmpty()) {
@@ -131,7 +131,7 @@ class IncomePieFragment : Fragment() {
                 val amount = incomeDataList[selectedMonthIndex][i].amount.toFloat()
                 val color = customColors[i % customColors.size]
                 categoriesMap[category] = Pair(categoriesMap.getOrDefault(category, Pair(0f, color)).first + amount, color)
-                totalForMonth += amount.toInt()
+                totalForMonth += amount.toDouble()
             }
         }
 
@@ -158,7 +158,7 @@ class IncomePieFragment : Fragment() {
         binding.pieChart.description.isEnabled = false
         binding.pieChart.legend.isEnabled = false
         binding.pieChart.setCenterTextSize(24f)
-        if (totalForMonth != 0) {
+        if (totalForMonth != 0.0) {
             binding.pieChart.centerText = totalForMonth.toString()
         }
         binding.pieChart.animateY(1000)
